requirements.txt
Flask
langchain
openai
python-dotenv
langchain_community
fitz
frontend
reportlab